Fixed warnings : format specifier to int from long in printf (use %ld) ; comparison signed/unsigned; conversion from double to int

+++ Mobility/SingletonRandomPattern.h
@@ -50,7 +50,7 @@ private:
   static int refCount;
   static int currentRefCounter;
 
-  int _x, _y;
+  double _x, _y;
   double _nextInterval;
 
   static RandomPattern* _instance;
