+++ NetworkInterfaces/Wireless/WESingleQueue.h
@@ -47,7 +47,7 @@ public:
     virtual int getCollided(int);
     virtual double getProbTxInSlot(int, double, WirelessEtherModule*) { return 0; }
     virtual double getLambda(int) { return 1; }
-    virtual int getAvgFrameSize(int);
+    virtual double getAvgFrameSize(int);
     virtual int size();
     virtual int sizeOfQ(int) { return size(); }
